abstract out shared loading trigger logic
